+++ src/index.js
@@ -4,6 +4,7 @@ import './helpers/intercom';
 import { injectGlobal } from 'styled-components';
 import { globalStyles } from './styles';
 import { bootIntercom } from './helpers/utilities';
+import lang, { resources } from './languages';
 import Root from './Root';
 
 // eslint-disable-next-line
@@ -11,5 +12,12 @@ injectGlobal`${globalStyles}`;
 
 // Intercom
 bootIntercom();
+
+// Languages (i18n)
+lang.init({
+  lng: 'en',
+  debug: process.env.NODE_ENV === 'development',
+  resources,
+});
 
 ReactDOM.render(<Root />, document.getElementById('root'));